Improve speed by using compiled regular expressions
Although regular expressions were compiled before using them, the compiled
expressions were not used (in fact, they were converted back to a string
and then compiled again). This fixes that and results in a performance
improvement in particular when many target file names are given on the
command line. For 150 targets, running time for `snakemake -n` dropped from
97s to 37s.