Shadow directory support on a rule by rule basis
The new keyword shadow, with possible values 'shallow', 'full', or
simply True (equivalent to 'full'), create a shadowing directory that
mirrors the cwd with symlinks. With 'shallow', it only mirrors the
current directory, but 'full' recursively symlinks every file in cwd
in a shadowing directory tree.

The shadowed directory tree is in .snakemake/shadow, and is
automatically cleaned up on success and output files moved to their
correct locations. This helps deal with programs that generate extra
garbage files so that they can be cleaned up without doing it manually
in a shell command or via temp declarations. It also helps keep the cwd
clean during program runtime.

It simplifes shell command development because each job is in an
isolated cwd, so there's no need to prefix filenames for parallel jobs
with each sample to avoid collisions (or unintentional collisions
between programs or invocations of Snakemake).

Obviously, this won't work well if your Snakefile rule references ../
paths, but that shouldn't be common.