Initial commit with support for remote files.
This introduces changes that scratch an itch within our research group for remote file support, and I hope the new functionality will be useful for others as well.

What is it all about? Wouldn't it be wonderful if you would use Snakemake with remote data stored on Amazon S3? Imagine being able to read and write the same files as your collaborators without rsyncing files back and forth or using pricey network-attached storage. Want to output and store 1PB of small files? Want to let someone else worry about durability? Want to work with large files locally that you can store individually but that are too large in aggregate? The support for remote storage introduced with these changes will allow you do all of these things! Remote storage is initially intended to work with AWS S3 and things sporting a similar API, though it is flexible enough that would could extended it to support other remote providers. It expect you to already have an S3 bucket setup, and API credentials for it permitting read/write access.

Here's how you use it:

The `Snakefile` now supports a wrapper function, `remote()`, indicating a file is on a remote storage provider (this is similar to `temp()` or `protected()`). During rule execution, the file (or object) specified, is downloaded to the `cwd`, within a sub-directory bearing the same name as the bucket. This lets you have multiple remote origins with reduced likelihood of name collisions, and allows snakemake to easily translate remote objects to local file paths. The `remote()` wrapper is mutually-exclusive with the `temp()` and `protected()` wrappers.

The implementation of all communication with the remote storage provider is specified in a class passed as a keyword argument to the remote wrapper. This class defines methods related to being a RemoteObject(). By default, the S3 provider is used and it does not need to be specified explicitly. If you are interested in creating a custom remote storage provider for whatever system you are using, implement the methods defined in the abstract base class `snakemake.remote_providers.RemoteObjectProvider`. Take a look at the S3 provider for an example implementation. To use a custom remote storage provider implementation (FTP, etc.), import it within your `Snakefile`, and set the `provider` keyword argument, as is being done with S3 below:

    import snakemake.remote_providers.S3 as S3

    rule all:
        input:
            remote("bucket-name/file.txt", provider=S3)

Since S3 is the default remote storage provider, it can be omitted from the `Snakefile`:

    rule all:
        input:
            remote("bucket-name/file.txt")

Additional kwargs can be passed to the remote provider this way to set credentials, should you not want to rely on whichever credentials happen to be stored in the environment variables (suggestion: if using multiple buckets with different credentials, read in the key and secret within your `Snakefile` and pass as appropriate to the `remote()` wrappers). For the S3 provider, the additional kwargs are passed straight through to the boto `connect_s3()` method, so refer to the [boto docs](https://boto.readthedocs.org/en/latest/ref/s3.html#boto.s3.connection.S3Connection) to see the available parameters:

    rule all:
        input:
            remote("bucket-name/file.txt", provider=S3, additional_kwargs={"aws_access_key_id":"myKey", "aws_secret_access_key":"mySecret"})

This can also be used to specify a custom URL to connect to rather than AWS, for a different service offering an S3-like API:

    rule all:
        input:
            remote("bucket-name/file.txt", provider=S3, additional_kwargs={"host":"mystorageapi.example.com"})

For any storage backend with an API deviating significantly from that of S3, it may make sense to create a new remote provider.

Expand still works as expected, just wrap the expansion:

    rule all:
        input:
            remote(expand("bucket-name/{letter}-2.txt", letter=["A", "B", "C"]), provider=S3)

Only remote files needed to satisfy the DAG build are downloaded for the workflow.  By default, remote files downloaded prior to rule execution and are removed locally as soon as no rules depend on them. Remote files can be explicitly kept by setting the `keep=True` keyword argument:

    rule all:
        input: remote('bucket-name/prefix{split_id}.txt', keep=True)

If you wish to have a rule to simply download a file to a local copy, you can do so by declaring the same file path locally as is used by the remote file:

    rule all:
        input:
            remote("bucket-name/out.txt")
        output:
            "bucket-name/out.txt"
        run:
            shell("cp {output[0]} ./")

Here's a brief summary of the changes. I tried to minimize the impact on existing code, relying where possible on external classes, new functions, and the use of decorators to connect them.

Within `io.py`, the usage of flags has been extended to store not just boolean values, but also objects. This is how, for example, the remote provider is passed to the `_IOFile()` object.

Several methods of `_IOFile()` now bear a decorator "@_referToRemote". When called, this decorator ensures that for files flagged as `remote()`, the methods defined in the class for the remote provider are used rather than the local equivalents specified within the `_IOFile()` class of `io.py`.

The `dag.py` file has been updated to have a `handle_remote()` method that is responsible for handling remote files after job execution. It uploads files that are not remote, and removes local files flagged as remote when they are no longer needed (exception: if the `"keep=True"` argument is provided to `remote()`).

A `decorators.py` file has been added for general-purpose decorators. Currently in use is `@decAllMethods`, which is a class decorator that applies a specified decorator to all methods of a decorated class. This is used with in the `test_remote/` unit test to help mock out methods of `RemoteObject()`.

Two new exceptions have been added, `RemoteFileException()` and `S3FileException()`. These could probably be used more.

The `jobs.py` file has been changed to add methods related to assessing remote files for a given job. It also has two new functions used in `dag.py` `handle_remote()` to delete remote-containing directories when they are empty after purging local copies of remote files. It also adds `expanded_input()`, which is used in `handle_remote()` to help determine which wildcard/dynamic files should be removed after job execution. The function `expanded_input()` would benefit from a critical eye.

The `rules.py` file has a small change causing newly expanded `_IOFile`s to clone the flags of their unexpanded sources. There's probably a better way.

The `workflow.py` file has been modified to import the new `glob_wildcards_remote()` helper function from `io.py`. This function is the remote analog to `glob_wildcards()`, but it needs to know the remote provider (S3 is default).

A unit test has been added, `test_remote/`. This is similar in nature to the function `test_cluster_dynamic/`. For testing, the `moto` package is used to mock out boto, so that remote file tests do not need to hit the live S3 API.

These changes have the following dependencies (as represented in setup.py):
* boto>=2.38.0 (for making AWS/S3 API calls)
* filechunkio>=1.6 (for file chunking prior to multipart upload, avoiding wheel reinvention)
* moto>=0.4.14 (for mocking up local S3 "buckets" for unit testing)

This is a first pass at support for remote storage support, and feedback is welcome. The intent was to add remote file support while make minimal changes to existing Snakemake logic.