Don't abort when a DRMAA job fails to terminate
Previously, if any job failed to cancel, Snakemake would abort without
attempting to cancel the remaining jobs, leaving them running. This
could happen if a job changed it status to "completed" between the last
status update and the cancel attempt, in which case Snakemake would
attempt to cancel a completed job, throwing an InvalidJobException. The
new code ignores this exception and keeps going, ensuring that Snakemake
will try to cancel every job.