Fix test_static_remote under nosetests
Had to rename the S3Mocked module and the bucket used in the test to
avoid collisions with a module and bucket of the same names in another
test. For some reason the import and the bucket contents are persistant
between tests when running "python setyup.py nosetests" but not when
running "python setup.py test".