Make "python3 -m snakemake" work
If necessary for any reason, this can make it very clear which particular
Python interpreter is used to run a snakemake instance. For example,
there might be multiple virtualenvs with snakemake and different external
packages installed. This would be one way to choose the correct snakemake.

For a discussion, see also https://github.com/pypa/pip/issues/3164 .