Add a failing API test
It fails because the snakemakepath parameter is left at its default value of
None, leading to an error when snakemake tries to re-execute itself. The
problem did not appear in previous versions because the code in "run:"
directives was executed in-process while it is now executed in separate
processes.

The error message on failure is:
/bin/sh: 2: None: not found