Fix Namedlist.allitems()
If the last item in _names it iterates over is an empty list, then it
would yield one item too many. Since iteration order varies from run to
run, this problem would not always occur. It also manifested itself only
because Snakemake 3.9.1 started to check whether there are duplicate
outputs. Since the extra item was yielded twice (once with a name, once
without), it would trigger the error.

The fix is to modify _names attribute: Previously, it would map
a name to a (start, end) tuple, where the end could also be None, which
would be seen as equivalent to start + 1. The problem was that these
intervals were sorted in allitems(), but only by start coordinate, which
would sometimes sort (3, None) before (3, 3). We now make the end explicit
(not allowing None as end index) and use it as a tie breaker. Thus, the
order becomes (3, 3), (3, 4) and we get a correct result.

Fixes #574