Consider temp file size when scheduling.
Usually, this should guide the scheduler to first schedule all jobs
that remove the largest temp file, then the second largest and so on.
Since the weight is summed up, it can in theory be that it sometimes
prefers a set of many jobs that all depend on smaller temp files though.

A real solution to the problem is therefore to use dummy jobs.