Fully enable to use both singularity and conda for the same rule. Further, allow to define singularity/docker image globally.
This will create the conda env from within the container, which
offers the nice sweet spot to freely choose the level of reproducibility with
snakemake.
1. conda based software deployment, 2. singularity based software deployment,
3. conda for software definition, singularity for OS definition.

While conda envs will be created and loaded within the container, they are
persisted to the working directory, with a hash combining both the singularity image and the env definition.