Add option to the shell class to use bash on Windows
This fix makes it possible to use bash as an external executable on Win.
It also adds a  shell.use_bash_on_win(bashcmd) convenience function to easily
configure bash.
If bashcmd is None it will try to locate a bash from either a conda installed
git or from system installed "git for windows"