Allow multiple configfiles to be specified
Changed the specification of `--configfile` to use action 'append' to
specify configfile as a list of configfiles (one at a time) instead of a
single path. Updated creation of overwriting configuration to iterate
through the configuration file list one at a time, updating as
necessary. Updated executors to correctly specify the new list of
configfiles in job scripts. This allows for multiple configuration files
to be used to update the default configuration file specified in a
Snakefile, allowing more flexible modification of default parameters.