Merged in SamStudio8/snakemake (pull request #370)
Prevent flattening lists inside rule input dict

Moves setting of `is_iterable` flag inside `_apply_wildcards` to
ensure that it is correctly reset to True, if the current `item`
in `olditems` is a dictionary. This ensures that each element of
a dict returned by a user function via `unpack` is correctly
checked for iterability.