adding crc32c checksum for Google cloud storage (#273)
The Google Storage api does not expose a function to stream to file,
however the blob.write_to_file accepts a file object, so we can create
a class to wrap this object, exposing the same functions, but also
generating a hasher that we can update for each write action. This allows
us to calculate the crc32c while the file is being downloaded and written
to file as opposed to needing to download and then read again, which will
make a difference for large files. See the docstrings in the GS.py remote
file for links to further documentation

Signed-off-by: vsoch <vsochat@stanford.edu>

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>