Add output redirect keywords.
Workflows involving many shell scripts can become quite cluttered
with bash redirections if every line in the script has them, especially
if stdout and stderr are redirected to separate files.

This allows the developer to instruct Snakemake to redirect the entirety
of a given script over to files, by flagging log files with stdout and
stderr.