Developers can display specific log files on fail.
The --show-failed-logs is useful but can be overly
expressive.  For example, on some of my workflows,
there may be 10000 lines of log output in five files,
useful for troubleshooting the outcomes of calculations
that did not end in error codes.

But just one of those files, a short one, contains the
critical error info.

If just that file prints out during a failure, my operators
will most of the time immediately be able to see what has
failed.