only assert statinfo is identical for matches
XRootD will occasionally return multiple matches for a file. This behavior is transient and probably harmless, but it causes the assertion in _get_statinfo to fail. This updates the assertion to check only the relevant properties. Note that, even with all relevant properties identical, match[0].statinfo != match[1].statinfo, hence the wordiness of the assertion.