adding support for google-lifesciences gpu
To add support for GPU, the user is limited to n1- general machine
types, and we do this by way of setting the self._machine_type_prefix
to be n1 (if it does not already start with n1). We can then choose
an accelerator type that is greater than or equal to what the user has
requested, and then choose from that set. With the current implementation
the P100 type will be the first chosen, and this makes sense as (I think)
it is the cheapest GPU option, something likely wanted by researchers
running these at scale. In the future (when someone asks for it) we can
add some flag that will allow the user to further specify a GPU accelerator
type, but for now this seems reasonable. Once the accelerator is added
to the VM, we can confirm that it creates an instance with GPU support

Signed-off-by: vsoch <vsochat@stanford.edu>