snakefile in tests/common.py should switch between tempdir and original
currently, some tests use the snakefile in the temporary working
directory, and others user the original. There is a flag, no_tmpdir,
that can distinguish this for the working directory specification,
and the same needs to be done for the snakefile, as different
tests have different expectations for its location.

Signed-off-by: vsoch <vsochat@stanford.edu>