Google Life Sciences executor must use relative paths for build package
Currently, if the user puts a path for a snakefile that is not relative to
the working directory where the build package is generated, the replace function
(that replaces the workdir in each file path) will not work and the build package
will be generated with a hard coded path on the user system. Since our options are
to try and muck around with moving files, or hackily changing paths for the included
files (not ideal) a much cleaner solution is to do similar to what a docker build does,
and require context to be at or for some subfolder level of the working directory.
Most runs that Ive seen will be for a Snakefile in a GitHub repo, meaning the working
directory includes the snakefile, so I think this will be a rare use case. I still
want to catch it, in case it does occur!

Signed-off-by: vsoch <vsochat@stanford.edu>