adding machine and gpu_model job resources
this will add two optional job resources to the Google Life Sciences executor,
including a machine (to mirror --machine-type-prefix) and a gpu_model (for
the user to select from the cloud or vendor specific namespace of gpu models,
e.g., nvidia-tesla-p4. If the user specifies a gpu_model without gpu or nvidia_gpu,
the count will default to 1. The --machine-type-prefix command line arg also
overrides the job resource, since this could be desired. The docs are also
updated with these changes.

Signed-off-by: vsoch <vsochat@stanford.edu>