adding work in progress google lifesciences tutorial
This PR is a work in progress to add a tutorial for using
the Google Life Sciences executor. There are some bugs currently
with running and inspecting jobs for the API, so I am pausing
work until these are resolved. The PR also adds a new group
for executor tutorials like this one, and references
a helper file to upload data to Cloud Storage from a client

Signed-off-by: vsoch <vsochat@stanford.edu>