🐛 Fix bug to work with git submodules
When working with Kubernetes, files returned by "git ls-files" will be
deemed as source files and then uploaded to Kubernetes pods as secrets.

However, if the working directory is a git repository with submodules,
the submodule directory will be returned by "git ls-files" as well, but
it's a directory rather than a regular file, so previously snakemake
may fail.

The bug can be fixed simply by adding "--recurse-submodules" to the git
command.