Adding work in progress google lifesciences tutorial (#331)
* adding work in progress google lifesciences tutorial

This PR is a work in progress to add a tutorial for using
the Google Life Sciences executor. There are some bugs currently
with running and inspecting jobs for the API, so I am pausing
work until these are resolved. The PR also adds a new group
for executor tutorials like this one, and references
a helper file to upload data to Cloud Storage from a client

Signed-off-by: vsoch <vsochat@stanford.edu>

* adding an error and debugging to finish up google executor tutorial

Signed-off-by: vsoch <vsochat@stanford.edu>

* Updating Snakemake version to be >=5.16

Co-Authored-By: Johannes Köster <johannes.koester@uni-due.de>

* fixes to google life sciences executor and addition of log file section

Signed-off-by: vsoch <vsochat@stanford.edu>

* Update google_lifesciences.rst

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>