use multi-stage builds + install singularity from source + use nonroot user
The Dockerfile prior to this commit install Singularity version 2.x. This commit installs Singularity 3.x from source. This commit also adds multi-stage builds, so unnecessary programs are not present in the final image. The separate stages can be built in parallel with `DOCKER_BUILDKIT=1 docker build .` on systems that have BuildKit.

This commit is a result of the comment https://github.com/snakemake/snakemake/pull/371#issuecomment-625167668. Podman can replace the Singularity installation in the future.