more work on adding container wrapper
we now have a deployment docker class that can parse a container URI, and determine i
f it exists or not. The snakemake/wrapper.py is now transformed into a class, but
some of the previous functionality is not clear to me (e.g., what a git+file means)
so I did my best to start writing the refactored class. It would be best to simplify
it more, but I cannot do that yet because I dont have an understanding of all the kinds
of strings that can be specified for a wrapper in the Snakefile

Signed-off-by: vsoch <vsochat@stanford.edu>