code no longer crashes on report for run rules
This commit fixes a bug in v5.26.* that was causing `snakemake --report` to crash when there were `run` rules.

The code has never handled this well: it seems like the desired behavior would be to make the "source" in the report for `run` rules to actually be the Python code being run. In < v5.26, instead in just printed "source" instead of the Python code. But then in v5.26.*, it was changed so that the code was trying to iterate over a `None` object, so the `snakemake --report` directive actually crashed when there was a `run` rule. This commit makes it go back to doing the old behavior of just printing "source", so at least it doesn't crash even if the report is uninformative as to the code being run.

Ideally, someone who knows how to actually access the code for `run` cells could fix this better.

Addresses issue #682.