Ensure output files of failed jobs are always removed in cluster mode. (#857)
Originally, when running in `--cluster` mode, snakemake relied on the child snakemake-processes within the cluster jobs to delete their output files if they failed. In practice, e.g. when using the SLURM scheduler, the jobs are killed with SIGKILL if they run over their time or memory limits. This made it impossible for them to clean up their own output files. There are workarounds for that, but especially running out of memory seems impossible to handle to me.

That is why the master process of snakemake should handle the cleanup, if running on a cluster. This process can detect if a job has failed or not (with custom `--cluster-status` if necessary), and can then clean up after the child process.

This works at least if the master process is active during the whole computation, but e.g. not with `--immediate-submit`.

This commit also adds a test case (test_github_issue627) that simulates a cluster environment. It launches a child job and kills it with SIGKILL, and asserts that
1. snakemake terminates with exit code != 0 and
2. the output file of that child job is deleted.

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>