Updated the rule for variant calling.
The rule is currently using two utilities for making the variant
calling. Those utilities are samtools and bcftools.
While following the tutorial using both utilities, a warning showed up,
claiming that parameter -g of samtools mpileup is deprecated. After
researching for a bit I found out that the command bcftools is
recommended in order to avoid errors resulting from use of incompatible
versions of samtools and bcftools when using in the mpileup+bcftools
call pipeline.

-What does option -O do?
It dictates output type.

-What means the 'v' argument for option -O?
It set the output format to uncompressed VCF