Fix .jobfinished permissions on NFS
A file system delay (?) can cause the temporary files, like `.jobfinished`
to be present but not writable on NFS system. In this case, this patch
catches the error, forces an update of the permissions, and tries to
remove the file again. Since there is no reason this issue should exist
only for `.jobfinished` files, the same changes were implemented for
`.jobscript` and `.jobfailed`.