updating with testing of spack copy/relocate environments
It was just added to spack to be able to relocate/copy entire views, but I do ont
think this will work because the paths are consistently too long given the .snakemake
metadata folder, spack folders and then the hash directories. So it would not be reasonable
to support because most cases where the user is working from a nested folder on HPC it just
will not work. However using spack views as base environments I think is still useful.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>