removing spack archive for now
spack was updated so that copy is supported, but the issue for archive is that the
view path must be shorter than the install path, which is hugely unlikely for a snakemake
archive that is very nested. So for the time being it is best to disable
creating archives of spack environments, and this can be re-visited when someone
uses it, requests it, and there is further discussion.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>