Create the base shadow directory per job.
Originally, the base shadow directory was created by the main snakemake process, however this prevents to use a node-local shadow directory in clusters.
Now the creation of the base shadow directory is allowed to fail in `persistency.py`, and each job ensures that the base shadow directory exists in `jobs.py`.

This commit also includes a test case and modified documentation.