fix: quote jobid passed to status script to support multi-cluster Slurm setup (#1459)
* fix: quote jobid passed to status script to support multi-cluster setup

Motivation: In a multi-cluster Slurm setup, i.e. passing the flag
`--clusters` to `sbatch`, then the job id returned with the `--parsable`
flag is actually `jobid;cluster_name`. This breaks a custom status
script because the shell ends the command at the semi-colon. This PR
quotes `jobid;cluster_name` so that a custom status script could then
parse this input to obtain the job status from the specified
cluster_name and jobid.

* Add a test for multi-cluster status script

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>