fix: issue1460 intermediate files existing (or not) (#1541)
* chore: add testcase for issue #1460

* fix expected results

* try force update of inventory after upload

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* run gls tests first

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* pushing test fix

the issue is that the blob object held by a GS remote object can go sort
of stale, returning a False for blob.exists() when it clearly exists! To fix
we need to do an additional self.update_blob() and then returning the exists
check again. I am not sure if this can be made more efficient by only checking
under certain conditions, but since it seems likely we cannot perfectly know when
the blob has gone stale the sure way is to always update.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* wrong version of black

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* remove unused remote inventory, was just testing!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>