Refactor group job calculation into new class
Moves all relevant code into the GroupResources class in the resources
module.

Remove the hardcoding of "runtime" from the logic and isolate it into
two variables: additive_resources, which determines which resources are
summed across layers, and sortby, which determines sorting priority when
layering.

GroupResources is changed a UserDict containing a class method
'basic_layered'. This opens the possibility for new group resource
merging methods to be added later.