fix: Handle temp files for all jobs in a group.
Fixes #1754 and by extension #1732. See #1754 specifically for
a detailed analysis of the problem. Essentially, the issue was
that the `job` argument to `dag.DAG.finish()` was being reused as
a local variable inside the method. So when `job` was group of jobs,
it would only handle temp files for the last job in that group.
This fixes that by calling `handle_temp()` explicitly for each
individual job. (Note that `handle_temp()` would also know how to
deal with a job group.)