feat: adding flux executor (#1810)
Signed-off-by: vsoch <vsoch@users.noreply.github.com>

### Description

This will add the Flux Executor to snakemake. Flux is a very
programmatic job manager (maybe that's too limited a term) but it can be
run in the context of HPC or cloud! I've written a containerized
tutorial so reviewers can test it out. Also pinging @trws and @grondo to
take a look.

@johanneskoester since this requires a new executor backend, let me know
how you'd like to test. We do have a container so we could try a
strategy that uses that.

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>