fix: Add back logging of run directives (#1883)
### Description

Fix #1698 exception tracebacks in run: blocks are sent to /dev/null,
making rules hard to debug

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


I have created a text based on #1881 but I don't see how I can test that
the error message is printed out correctly to a log.