fix: fix `--immediate-submit` (#1851)
### Description

parameter `--immediate-submit` is expected to [submit all jobs to the
cluster instead of waiting for present input
files](https://snakemake.readthedocs.io/en/stable/executing/cli.html?highlight=immediate#CLUSTER).
However, as tested in the system, nothing changed compared to "run
without this flag".

Detailed discussions can be found in #1865 and #1715. Briefly, it is
caused by a hotfix for googlelifescience API handling.

### QC
* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] ~The documentation (`docs/`) is updated to reflect the changes~ or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).