fix: Comparison of rules and non-rule instances (#1894)
### Description

Fixes #1891 and #1892. The issue is that the [`__eq__` method of
`Rule`](https://github.com/snakemake/snakemake/blob/274d613fd27b632d5bbe4c6fe55a82262d6b2dec/snakemake/rules.py#L1221)
implicitly expects `other` to be an instance of `Rule` as well. Thus,
any comparison of a `Rule` instance to any non-`Rule` object leads to an
`AttributeError`. This is a more general fix than #1893.

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).