feat: k8s: add --k8s-cpu-scalar (#1857)
### Description

see #1856. In short, this adds the ability to scale CPU requests to most
efficiently use the CPUs of underlying nodes of a k8s cluster, of which
typically a small fraction are reserved for system pods.

Open questions for maintainers:

- Is this a reasonable approach?
- Does this need to be extended to other resource requests (i.e.
memory)?

### QC

(these aren't done yet, I'll do them once a maintainer approves of the
design of this solution)

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behaviour or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>