fix: allow report generation to handle pathlib objects (#1904)
### Description

In #1869 I added the ability to use a `pathlib.Path` object in the
`script` directive. When generating a report where a rule has a Path in
the script directory, this function was causing snakemake to error as it
assumes the path is a string. This PR makes it explicitly a string.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).